tetragon/windows: Compilation only Change for pkg/procsyms on Windows
Stubbing out the GetFnSymbol function on Windows to make sure pkg/procsysms compiles on Windows.
